Refactor the EditorProvider component and extract hooks
